Fix: DTO

+++ velo/src/main/java/com/velo/route/domain/Route.java
@@ -3,5 +3,5 @@ package com.velo.route.domain;
 public record Route(
     String polyline,
     double distance,
-    int time
+    long time
 ) {}
