+++ velo/src/main/java/com/velo/route/dto/RouteDTO.java
@@ -3,5 +3,5 @@ package com.velo.route.dto;
 public record RouteDTO(
     String polyline,
     double distance,
-    int time
+    long time
 ) {}
